Create Letters-in-Organic-Chemistry
https://benthamscience.com/journals/letters-in-organic-chemistry/
https://benthamscience.com/journals/letters-in-organic-chemistry/author-guidelines/
REFERENCES
References must be listed in the ACS Style only. All references should be numbered sequentially [in square brackets] in the text and listed in the same numerical order in the reference section. The reference numbers must be finalized and the bibliography must be fully formatted before submission.

See below few examples of references listed in the ACS Style

Journal Reference
[1]  Bard, M.; Woods, R.A.; Bartón, D.H.; Corrie, J.E.; Widdowson, D.A. Lipids, 1977, 12, 645-654.
[2]  Zhang, W.; Brombosz, S.M.; Mendoza, J.L.; Moore, J.S. J. Org. Chem., 2005, 70, 10198-10201.
Book Reference
[3]  Crabtree, R.H. The Organometallic Chemistry of the Transition Metals, 3rd ed.; Wiley & Sons: New York, 2001.
Book Chapter Reference
[4]  Wheeler, D.M.S.; Wheeler, M.M. D. Stereoselective Syntheses of Doxorubicin and Related Compounds In: Studies in Natural Products Chemistry; Atta-ur-Rahman, Ed.; Elsevier Science B. V: Amsterdam, 1994; Vol. 14, pp. 3-46.
Conference Proceedings
[5]  Jakeman, D.L.; Withers, S.G.E. In: Carbohydrate Bioengineering: Interdisciplinary Approaches, Proceedings of the 4th Carbohydrate Bioengineering Meeting, Stockholm, Sweden, June 10-13, 2001; Teeri, T.T.; Svensson, B.; Gilbert, H.J.; Feizi, T., Eds.; Royal Society of Chemistry: Cambridge, UK, 2002; pp. 3-8.
URL(WebPage)
[6]  National Library of Medicine. Specialized Information Services: Toxicology and Environmental Health. sis.nlm.ni0068.gov/Tox/ToxMain.html (Accessed May 23, 2004).
Patent
[7]  Hoch, J.A.; Huang, S. Screening methods for the identification of novel antibiotics. U.S. Patent 6,043,045, March 28, 2000.
Thesis
[8]  Mackel, H. Capturing the Spectra of Silicon Solar Cells. PhD Thesis, The Australian National University: Canberra, December 2004.
E-citations
[9]  Citations for articles/material published exclusively online or in open access (free-to-view), must contain the accurate Web addresses (URLs) at the end of the reference(s), except those posted on an author’s Web site (unless editorially essential), e.g. ‘Reference: Available from: URL’.
Some important points to remember

All references must be complete and accurate.
All authors must be cited and there should be no use of the phrase et al.
Date of access should be provided for online citations.
Journal names should be abbreviated according to the Index Medicus/MEDLINE.
Punctuation should be properly applied as mentioned in the examples given above.
Superscript in the in-text citations and reference section should be avoided.
Abstracts, unpublished data and personal communications (which can only be included if prior permission has been obtained) should not be given in the references section. The details may however appear in the footnotes.
The authors are encouraged to use a recent version of EndNote (version 5 and above) or Reference Manager (version 10) when formatting their reference list, as this allows references to be automatically extracted.